refactor : OmokRepository 클래스 생성
